pyproject.toml
[project]
name = "qwen"
version = "0.1.0"
description = "Qwen3.5-Only Virtual Try-On Pipeline - Simple virtual try-on using only Qwen3.5 multimodal model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]
